Tests/LibWeb/Text/input/abspos-relayout-insert-fixed-fallback.html

<!DOCTYPE html>
<script src="include.js"></script>
<style>
    body { margin: 0; }
    #root { position: relative; width: 360px; height: 180px; }
    #boundary {
        position: absolute;
        left: 40px;
        top: 30px;
        width: 160px;
        height: 90px;
        overflow: hidden;
    }
    #inner { width: 80px; height: 30px; }
    #fixed {
        position: fixed;
        left: 25px;
        top: 18px;
        width: 35px;
        height: 20px;
    }
</style>
<div id="root">
    <div id="boundary">
        <div id="inner"></div>
    </div>
</div>
<script>
    function rectToString(rect) {
        return `${rect.left},${rect.top},${rect.width},${rect.height}`;
    }

    asyncTest((done) => {
        document.body.offsetHeight;

        const partialBefore = internals.partialLayoutCount();
        const fullBefore = internals.fullLayoutCount();

        const fixed = document.createElement("div");
        fixed.id = "fixed";
        document.getElementById("inner").appendChild(fixed);
        document.body.offsetHeight;

        const fixedRect = fixed.getBoundingClientRect();
        const partialDelta = internals.partialLayoutCount() - partialBefore;
        const fullDelta = internals.fullLayoutCount() - fullBefore;

        println(`fixed=${rectToString(fixedRect)}`);
        println(`partial-layouts=${partialDelta} full-layouts=${fullDelta}`);
        println(rectToString(fixedRect) === "25,18,35,20"
            && partialDelta === 0 && fullDelta === 1 ? "PASS" : "FAIL");
        done();
    });
</script>
