Tests/LibWeb/Text/expected/abspos-relayout-insert-fixed-fallback.txt

fixed=25,18,35,20
partial-layouts=0 full-layouts=1
PASS
